refactor: update import paths for State and StateType to core.utils.state_base feat: implement Choice class for conditional expression evaluation refactor: add condition parsers for enhanced condition handling refactor: reorganize and implement condition parsing logic refactor: remove deprecated state_base module and migrate functionality
