+++ core/blocks/lambda_handler.py
@@ -3,7 +3,7 @@ from time import time
 from pathlib import Path
 import importlib.util
 
-from core.state_base import State, StateType
+from core.utils.state_base import State, StateType
 
 
 class Lambda(State):
