+++ core/blocks/parallel_handler.py
@@ -1,7 +1,7 @@
 from typing import Any, Optional
 import concurrent.futures
 
-from core.state_base import State, StateType
+from core.utils.state_base import State, StateType
 from core.state_machine import StateMachine
